+++ bin/format.dart
@@ -2,76 +2,23 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:path/path.dart' as path;
 
-import 'package:irdartfmt/src/dart_formatter.dart';
-import 'package:irdartfmt/src/exceptions.dart';
-import 'package:irdartfmt/src/formatter_options.dart';
+import 'package:irdartfmt/src/cli/formatter_options.dart';
+import 'package:irdartfmt/src/cli/options.dart';
+import 'package:irdartfmt/src/cli/output.dart';
+import 'package:irdartfmt/src/cli/show.dart';
+import 'package:irdartfmt/src/cli/summary.dart';
 import 'package:irdartfmt/src/io.dart';
-import 'package:irdartfmt/src/source_code.dart';
 import 'package:irdartfmt/src/style_fix.dart';
-
-// Note: The following line of code is modified by tool/grind.dart.
-const version = '1.3.3';
 
 void main(List<String> args) {
   var parser = ArgParser(allowTrailingOptions: true);
 
-  parser.addSeparator('Common options:');
-  parser.addFlag('help',
-      abbr: 'h', negatable: false, help: 'Shows usage information.');
-  parser.addFlag('version',
-      negatable: false, help: 'Shows version information.');
-  parser.addOption('line-length',
-      abbr: 'l', help: 'Wrap lines longer than this.', defaultsTo: '80');
-  parser.addFlag('overwrite',
-      abbr: 'w',
-      negatable: false,
-      help: 'Overwrite input files with formatted output.');
-  parser.addFlag('dry-run',
-      abbr: 'n',
-      negatable: false,
-      help: 'Show which files would be modified but make no changes.');
-
-  parser.addSeparator('Non-whitespace fixes (off by default):');
-  parser.addFlag('fix', negatable: false, help: 'Apply all style fixes.');
-
-  for (var fix in StyleFix.all) {
-    // TODO(rnystrom): Allow negating this if used in concert with "--fix"?
-    parser.addFlag('fix-${fix.name}', negatable: false, help: fix.description);
-  }
-
-  parser.addSeparator('Other options:');
-  parser.addOption('indent',
-      abbr: 'i', help: 'Spaces of leading indentation.', defaultsTo: '0');
-  parser.addFlag('machine',
-      abbr: 'm',
-      negatable: false,
-      help: 'Produce machine-readable JSON output.');
-  parser.addFlag('set-exit-if-changed',
-      negatable: false,
-      help: 'Return exit code 1 if there are any formatting changes.');
-  parser.addFlag('follow-links',
-      negatable: false,
-      help: 'Follow links to files and directories.\n'
-          'If unset, links will be ignored.');
-  parser.addOption('preserve',
-      help: 'Selection to preserve, formatted as "start:length".');
-  parser.addOption('stdin-name',
-      help: 'The path name to show when an error occurs in source read from '
-          'stdin.',
-      defaultsTo: '<stdin>');
-  parser.addOption('package-name',
-      help:
-          'The name of the package. Used to be able to sort imports with a package section.'
-          'The sort fix will not consider packages if not supplied.',
-      defaultsTo: null);
-  parser.addFlag('profile', negatable: false, hide: true);
-  parser.addFlag('transform', abbr: 't', negatable: false, hide: true);
+  defineOptions(parser, oldCli: true);
 
   ArgResults argResults;
   try {
@@ -86,23 +33,15 @@ void main(List<String> args) {
   }
 
   if (argResults['version']) {
-    print(version);
+    print(dartStyleVersion);
     return;
   }
 
-  // Can only preserve a selection when parsing from stdin.
   List<int> selection;
-  if (argResults['preserve'] != null && argResults.rest.isNotEmpty) {
-    usageError(parser, 'Can only use --preserve when reading from stdin.');
-  }
-
   try {
-    selection = parseSelection(argResults['preserve']);
-  } on FormatException catch (_) {
-    usageError(
-        parser,
-        '--preserve must be a colon-separated pair of integers, was '
-        '"${argResults['preserve']}".');
+    selection = parseSelection(argResults, 'preserve');
+  } on FormatException catch (exception) {
+    usageError(parser, exception.message);
   }
 
   if (argResults['dry-run'] && argResults['overwrite']) {
@@ -116,12 +55,16 @@ void main(List<String> args) {
     usageError(parser, 'Cannot use --$chosen and --$other at the same time.');
   }
 
-  var reporter = OutputReporter.print;
+  var show = Show.legacy;
+  var summary = Summary.none;
+  var output = Output.show;
+  var setExitIfChanged = false;
   if (argResults['dry-run']) {
     checkForReporterCollision('dry-run', 'overwrite');
     checkForReporterCollision('dry-run', 'machine');
 
-    reporter = OutputReporter.dryRun;
+    show = Show.dryRun;
+    output = Output.none;
   } else if (argResults['overwrite']) {
     checkForReporterCollision('overwrite', 'machine');
 
@@ -130,18 +73,15 @@ void main(List<String> args) {
           'Cannot use --overwrite without providing any paths to format.');
     }
 
-    reporter = OutputReporter.overwrite;
+    show = Show.overwrite;
+    output = Output.write;
   } else if (argResults['machine']) {
-    reporter = OutputReporter.printJson;
+    output = Output.json;
   }
 
-  if (argResults['profile']) {
-    reporter = ProfileReporter(reporter);
-  }
+  if (argResults['profile']) summary = Summary.profile();
 
-  if (argResults['set-exit-if-changed']) {
-    reporter = SetExitReporter(reporter);
-  }
+  setExitIfChanged = argResults['set-exit-if-changed'];
 
   int pageWidth;
   try {
@@ -189,12 +129,16 @@ void main(List<String> args) {
     // do nothing
   }
 
-  var options = FormatterOptions(reporter,
+  var options = FormatterOptions(
       indent: indent,
       pageWidth: pageWidth,
       followLinks: followLinks,
       fixes: fixes,
-      packageName: packageName);
+      packageName: packageName,
+      show: show,
+      output: output,
+      summary: summary,
+      setExitIfChanged: setExitIfChanged);
 
   if (argResults.rest.isEmpty) {
     formatStdin(options, selection, argResults['stdin-name'] as String);
@@ -202,83 +146,7 @@ void main(List<String> args) {
     formatPaths(options, argResults.rest);
   }
 
-  if (argResults['profile']) {
-    (reporter as ProfileReporter).showProfile();
-  }
-}
-
-List<int> parseSelection(String selection) {
-  if (selection == null) return null;
-
-  var coordinates = selection.split(':');
-  if (coordinates.length != 2) {
-    throw FormatException(
-        'Selection should be a colon-separated pair of integers, "123:45".');
-  }
-
-  return coordinates.map((coord) => coord.trim()).map(int.parse).toList();
-}
-
-/// Reads input from stdin until it's closed, and the formats it.
-void formatStdin(FormatterOptions options, List<int> selection, String name) {
-  var selectionStart = 0;
-  var selectionLength = 0;
-
-  if (selection != null) {
-    selectionStart = selection[0];
-    selectionLength = selection[1];
-  }
-
-  var input = StringBuffer();
-  stdin.transform(Utf8Decoder()).listen(input.write, onDone: () {
-    var formatter = DartFormatter(
-        indent: options.indent,
-        pageWidth: options.pageWidth,
-        fixes: options.fixes,
-        packageName: options.packageName);
-    try {
-      options.reporter.beforeFile(null, name);
-      var source = SourceCode(input.toString(),
-          uri: name,
-          selectionStart: selectionStart,
-          selectionLength: selectionLength);
-      var output = formatter.formatSource(source);
-      options.reporter
-          .afterFile(null, name, output, changed: source.text != output.text);
-      return;
-    } on FormatterException catch (err) {
-      stderr.writeln(err.message());
-      exitCode = 65; // sysexits.h: EX_DATAERR
-    } catch (err, stack) {
-      stderr.writeln('''Hit a bug in the formatter when formatting stdin.
-Please report at: github.com/dart-lang/irdartfmt/issues
-$err
-$stack''');
-      exitCode = 70; // sysexits.h: EX_SOFTWARE
-    }
-  });
-}
-
-/// Formats all of the files and directories given by [paths].
-void formatPaths(FormatterOptions options, List<String> paths) {
-  for (var path in paths) {
-    var directory = Directory(path);
-    if (directory.existsSync()) {
-      if (!processDirectory(options, directory)) {
-        exitCode = 65;
-      }
-      continue;
-    }
-
-    var file = File(path);
-    if (file.existsSync()) {
-      if (!processFile(options, file)) {
-        exitCode = 65;
-      }
-    } else {
-      stderr.writeln('No file or directory found at "$path".');
-    }
-  }
+  options.summary.show();
 }
 
 /// Prints [error] and usage help then exits with exit code 64.