+++ test/command_line_test.dart
@@ -13,11 +13,53 @@ import 'package:test/test.dart';
 import 'utils.dart';
 
 void main() {
-  test('exits with 0 on success', () async {
-    await d.dir('code', [d.file('a.dart', unformattedSource)]).create();
+  compileFormatterExecutable();
+
+  test('formats a directory', () async {
+    await d.dir('code', [
+      d.file('a.dart', unformattedSource),
+      d.file('b.dart', formattedSource),
+      d.file('c.dart', unformattedSource)
+    ]).create();
 
     var process = await runFormatterOnDir();
+    await expectLater(
+        process.stdout, emits(startsWith('Formatting directory')));
+
+    // Prints the formatting result.
+    await expectLater(process.stdout, emits(formattedOutput));
+    await expectLater(process.stdout, emits(formattedOutput));
+    await expectLater(process.stdout, emits(formattedOutput));
     await process.shouldExit(0);
+
+    // Does not overwrite by default.
+    await d.dir('code', [d.file('a.dart', unformattedSource)]).validate();
+    await d.dir('code', [d.file('c.dart', unformattedSource)]).validate();
+  });
+
+  test('Overwrites files', () async {
+    await d.dir('code', [
+      d.file('a.dart', unformattedSource),
+      d.file('b.dart', formattedSource),
+      d.file('c.dart', unformattedSource)
+    ]).create();
+
+    var process = await runFormatterOnDir(['--overwrite']);
+    await expectLater(
+        process.stdout, emits(startsWith('Formatting directory')));
+
+    // Prints whether each file was changed.
+    await expectLater(
+        process.stdout, emits('Formatted ${p.join('code', 'a.dart')}'));
+    await expectLater(
+        process.stdout, emits('Unchanged ${p.join('code', 'b.dart')}'));
+    await expectLater(
+        process.stdout, emits('Formatted ${p.join('code', 'c.dart')}'));
+    await process.shouldExit(0);
+
+    // Overwrites the files.
+    await d.dir('code', [d.file('a.dart', formattedSource)]).validate();
+    await d.dir('code', [d.file('c.dart', formattedSource)]).validate();
   });
 
   test('exits with 64 on a command line argument error', () async {
@@ -67,6 +109,16 @@ void main() {
 
     // Match something roughly semver-like.
     expect(await process.stdout.next, matches(RegExp(r'\d+\.\d+\.\d+.*')));
+    await process.shouldExit(0);
+  });
+
+  test('--help', () async {
+    var process = await runFormatter(['--help']);
+    await expectLater(
+        process.stdout, emits('Idiomatically formats Dart source code.'));
+    await expectLater(process.stdout, emits(''));
+    await expectLater(process.stdout,
+        emits('Usage:   dartfmt [options...] [files or directories...]'));
     await process.shouldExit(0);
   });
 
@@ -139,7 +191,6 @@ void main() {
       var process = await runFormatterOnDir(['--machine']);
 
       // The order isn't specified.
-
       expect(await process.stdout.next, anyOf(jsonA, jsonB));
       expect(await process.stdout.next, anyOf(jsonA, jsonB));
       await process.shouldExit();
@@ -171,7 +222,7 @@ void main() {
       await process.stdin.close();
 
       var json = jsonEncode({
-        'path': '<stdin>',
+        'path': 'stdin',
         'source': formattedSource,
         'selection': {'offset': 5, 'length': 9}
       });
@@ -199,7 +250,7 @@ void main() {
       var process = await runFormatter(['--indent', 'notanum']);
       await process.shouldExit(64);
 
-      process = await runFormatter(['--preserve', '-4']);
+      process = await runFormatter(['--indent', '-4']);
       await process.shouldExit(64);
     });
   });
@@ -303,19 +354,20 @@ void main() {
       await process.stdin.close();
 
       // No trailing newline at the end.
-      expect(await process.stdout.next, formattedSource.trimRight());
+      expect(await process.stdout.next, formattedOutput);
       await process.shouldExit(0);
     });
 
     test('allows specifying stdin path name', () async {
-      var process = await runFormatter(['--stdin-name=some/path.dart']);
+      var path = p.join('some', 'path.dart');
+      var process = await runFormatter(['--stdin-name=$path']);
       process.stdin.writeln('herp');
       await process.stdin.close();
 
       expect(await process.stderr.next,
           'Could not format because the source could not be parsed:');
       expect(await process.stderr.next, '');
-      expect(await process.stderr.next, contains('some/path.dart'));
+      expect(await process.stderr.next, contains(path));
       await process.stderr.cancel();
       await process.shouldExit(65);
     });